LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/zero-scroll-offset-expected.txt

abc

FAIL Scroll anchoring suppressed when scroll offset is zero. assert_equals: expected 0 but got 10
